+++ internal/impl/sql/util.go
@@ -2,14 +2,12 @@ package sql
 
 import (
 	"database/sql"
-
-	"github.com/Jeffail/benthos/v3/public/service"
 )
 
-func sqlResultJSONArrayCodec(rows *sql.Rows, msg *service.Message) error {
+func sqlRowsToArray(rows *sql.Rows) ([]interface{}, error) {
 	columnNames, err := rows.Columns()
 	if err != nil {
-		return err
+		return nil, err
 	}
 	jArray := []interface{}{}
 	for rows.Next() {
@@ -19,7 +17,7 @@ func sqlResultJSONArrayCodec(rows *sql.Rows, msg *service.Message) error {
 			valuesWrapped[i] = &values[i]
 		}
 		if err := rows.Scan(valuesWrapped...); err != nil {
-			return err
+			return nil, err
 		}
 		jObj := map[string]interface{}{}
 		for i, v := range values {
@@ -41,8 +39,40 @@ func sqlResultJSONArrayCodec(rows *sql.Rows, msg *service.Message) error {
 		jArray = append(jArray, jObj)
 	}
 	if err := rows.Err(); err != nil {
-		return err
+		return nil, err
 	}
-	msg.SetStructured(jArray)
-	return nil
+	return jArray, nil
+}
+
+func sqlRowToMap(rows *sql.Rows) (map[string]interface{}, error) {
+	columnNames, err := rows.Columns()
+	if err != nil {
+		return nil, err
+	}
+	values := make([]interface{}, len(columnNames))
+	valuesWrapped := make([]interface{}, len(columnNames))
+	for i := range values {
+		valuesWrapped[i] = &values[i]
+	}
+	if err := rows.Scan(valuesWrapped...); err != nil {
+		return nil, err
+	}
+	jObj := map[string]interface{}{}
+	for i, v := range values {
+		switch t := v.(type) {
+		case string:
+			jObj[columnNames[i]] = t
+		case []byte:
+			jObj[columnNames[i]] = string(t)
+		case int, int8, int16, int32, int64, uint, uint8, uint16, uint32, uint64:
+			jObj[columnNames[i]] = t
+		case float32, float64:
+			jObj[columnNames[i]] = t
+		case bool:
+			jObj[columnNames[i]] = t
+		default:
+			jObj[columnNames[i]] = t
+		}
+	}
+	return jObj, nil
 }