+++ CHANGELOG.md
@@ -11,7 +11,9 @@ All notable changes to this project will be documented in this file.
 - The `nats` input now supports metadata from headers when supported.
 - Field `headers` added to the `nats` output.
 - Go API: Optional field definitions added for config specs.
-- New (experimental) `sql_query` processor, which will supersede the existing `sql` processor for queries.
+- New (experimental) `sql_select` input.
+- New (experimental) `sql_select` and `sql_insert` processors, which will supersede the existing `sql` processor.
+- New (experimental) `sql_insert` output, which will supersede the existing `sql` output.
 
 ### Fixed